requirements.txt
streamlit
Pillow
PyMuPDF
pytesseract
python-docx
langchain
openai
pymupdf
langchain-community
streamlit-pdf-viewer
